Touched up documentation and ReadMe
