+++ README.md
@@ -1,30 +1,35 @@
 # test_job_scraper
 
 ### Running the project locally
-1.) clone the repo<br>
-2.) navigate to the project root<br>
-3.) activate your venv<br>
-4.) create your .env file
+1. clone the repo<br>
+2. navigate to the project root<br>
+3. activate and enter your venv<br>
+4. pip install -r requirements.txt
+5. create your .env file
 ```
 SECRET_KEY=[YOUR SECRET KEY HERE] or ask someone on the team for the secret key
-DATABASE_URL="postgresql:///job_crawler"
-DATABASE_NAME="job_crawler"
+DATABASE_URL=postgresql:///job_crawler
+TEST_DATABASE_URL=postgresql:///job_crawler
+DATABASE_NAME=job_crawler
 ```
-5.) run the following commands in the terminal to create your local copy of the database
+6. run the following commands in the terminal to create your local copy of the database
 ```
 psql
 CREATE DATABASE job_crawler
 ctrl+d to quit
 ```
-6.) in the terminal, run the `python3 migrate.py` this will drop all of the tables, create the latest version of the tables, and add the contents of the hrefs.csv file to your local database
+7. from the root project dir, populate your local database:
+```
+pg_dump -O database_utils/backup_database.sql | psql postgresql://job_crawler
+```
 
 
 ### Project Overview:
-- Who's it for?
+Who's it for?
 	- Bootcamp grads
-- What are we looking for?
+What are we looking for?
 	- Job descriptions that fit our needs (bootcamp grads, no degree)
-- Where are we looking for it?
+Where are we looking for it?
 	- Fortune 1000 company websites (for now) https://stillhiring.today/
 
 
@@ -40,19 +45,7 @@ ctrl+d to quit
 		5. careers.smartrecruiters.com: 7
 3. Scrape the job descriptions
 	- run them through OpenAI
-
-
-### Create a env file:
-
-
+4. Run the job descriptions through GPT
 
 ### REMINDER:
-
 Please `git pull` and `pip install -r requirements.txt` before you start working on the codebase to ensure you have the most up-to-date packages and code.
-
-
-## The Flow:
-get_job_boards.py > runs sqlite_db.py (creates job_boards.db)
-
-bulk_insert.py inserts job_boards.db into the working database
-
