+++ requirements.txt
@@ -1,3 +1,6 @@
+aiohttp==3.8.6
+aiosignal==1.3.1
+async-timeout==4.0.3
 attrs==23.1.0
 beautifulsoup4==4.12.2
 blinker==1.6.3
@@ -6,13 +9,18 @@ charset-normalizer==3.3.0
 click==8.1.7
 Flask==3.0.0
 Flask-SQLAlchemy==3.1.1
+frozenlist==1.4.0
+geonamescache==2.0.0
 h11==0.14.0
 idna==3.4
 itsdangerous==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.3
+multidict==6.0.4
+openai==0.28.1
 outcome==1.3.0
 psycopg==3.1.12
+psycopg2==2.9.9
 psycopg2-binary==2.9.9
 PySocks==1.7.1
 python-dotenv==1.0.0
@@ -21,11 +29,13 @@ selenium==4.14.0
 sniffio==1.3.0
 sortedcontainers==2.4.0
 soupsieve==2.5
-SQLAlchemy==2.0.22
+SQLAlchemy==2.0.23
 SQLAlchemy-Utils==0.41.1
+tqdm==4.66.1
 trio==0.22.2
 trio-websocket==0.11.1
 typing_extensions==4.8.0
 urllib3==2.0.7
 Werkzeug==3.0.0
 wsproto==1.2.0
+yarl==1.9.2
