+++ utilities/utils.py
@@ -12,8 +12,11 @@ load_dotenv()
 # connect to the local database directly for the insert
 # conn = psycopg2.connect(database=os.environ["DATABASE_NAME"])
 
-# Double checks DEV environment var and connects to appropriate DB
 def db_connect(DEV):
+    '''Double checks DEV environment var and connects to appropriate DB
+    TODO: this shouldn't live here.
+    '''
+
     if DEV:
         return psycopg2.connect(
             dbname=os.environ["DATABASE_NAME"]
@@ -235,9 +238,17 @@ def select_US_roles_entry():
         ])
         AND (json_response ->> 'location') NOT ILIKE '%India%' AND (json_response ->> 'location') NOT ILIKE '%latin%' AND (json_response ->> 'location') NOT ILIKE '%europe%' AND (json_response ->> 'location') NOT ILIKE '%UK%' AND (json_response ->> 'location') NOT ILIKE '%mexico%' AND (json_response ->> 'location') NOT ILIKE '%paris%'
         AND (job_title NOT ILIKE '%senior%' AND job_title NOT ILIKE '%staff%' AND job_title NOT ILIKE '%director%' AND job_title NOT ILIKE '%manager%' AND job_title NOT ILIKE '%sr.%' AND job_title NOT ILIKE '%data%' AND job_title NOT ILIKE '%head%' AND job_title NOT ILIKE '%sr %' AND job_title NOT ILIKE '%Mechanical%' AND job_title NOT ILIKE '%lead%' AND job_title NOT ILIKE '%net%' AND job_title NOT ILIKE '%Electrical%' AND job_title NOT ILIKE '%Principal%' AND job_title NOT ILIKE '%VP%' AND job_title NOT ILIKE '%Chassis%' AND job_title NOT ILIKE '%Legal%' AND job_title NOT ILIKE '%Avionics%' AND job_title NOT ILIKE '%President%')
-        AND NOT (json_response ? 'entry_level')
-        AND NOT job_url ILIKE 'https://jobs.lever.co/passportshipping%'
-        AND NOT job_url ILIKE 'https://jobs.lever.co/palantir%';
+        AND NOT (json_response::jsonb) ? 'apply';
     """
+    cursor.execute(select_query)
+    return cursor.fetchall()
+
+
+def select_applicable_jobs():
+    '''Query's job_postings and returns jobs given GPT's blessing '''
+
+    select_query = ''' SELECT * FROM job_postings WHERE (json_response::jsonb) ? 'apply' 
+    AND json_response ->> 'apply' ILIKE 'True';'''
+
     cursor.execute(select_query)
     return cursor.fetchall()
